Isolate hass state from base el (#3157)
