+++ src/external_app/external_events_forwarder.ts
@@ -1,7 +1,7 @@
 import { ExternalMessaging } from "./external_messaging";
 
 export const externalForwardConnectionEvents = (bus: ExternalMessaging) => {
-  document.addEventListener("connection-status", (ev) =>
+  window.addEventListener("connection-status", (ev) =>
     bus.fireMessage({
       type: "connection-status",
       payload: { event: ev.detail },
@@ -10,6 +10,6 @@ export const externalForwardConnectionEvents = (bus: ExternalMessaging) => {
 };
 
 export const externalForwardHaptics = (bus: ExternalMessaging) =>
-  document.addEventListener("haptic", (ev) =>
+  window.addEventListener("haptic", (ev) =>
     bus.fireMessage({ type: "haptic", payload: { hapticType: ev.detail } })
   );
