+++ src/resources/translations-metadata.ts
@@ -1,16 +1,5 @@
 import * as translationMetadata_ from "../../build-translations/translationMetadata.json";
-
-interface TranslationMetadata {
-  fragments: string[];
-  translations: {
-    [language: string]: {
-      nativeName: string;
-      fingerprints: {
-        [filename: string]: string;
-      };
-    };
-  };
-}
+import { TranslationMetadata } from "../types.js";
 
 export const translationMetadata = (translationMetadata_ as any)
   .default as TranslationMetadata;
