+++ src/state/more-info-mixin.ts
@@ -20,7 +20,7 @@ export default (superClass: Constructor<LitElement & HassBaseEl>) =>
       this.addEventListener("hass-more-info", (e) => this._handleMoreInfo(e));
 
       // Load it once we are having the initial rendering done.
-      import(/* webpackChunkName: "more-info-dialog" */ "../../dialogs/ha-more-info-dialog");
+      import(/* webpackChunkName: "more-info-dialog" */ "../dialogs/ha-more-info-dialog");
     }
 
     private async _handleMoreInfo(ev) {
